Import current development state
This project started as a simple Leaflet application ("I want a map that
is fullscreen, without stuff around it"), but developed to something
with (useful) features.

It is mostly a conglomerate of different leaflet addons, some with
custom modifications to make them compatible (I'll make these changes
visible in the next commits).